Big refactor.
Using callbacks in user code to react to matching an argument makes things very cumbersome. Instead, the API was refactored to return a ParsedArguments instance which makes traversing the parsed arguments simple in the 90% case, and flexible enough in the face of arguments that may appear multiple times, or when the order of arguments is important. See the README.md for a quick overview on the new API.